+++ feagi_data_serialization/src/byte_structure/mod.rs
@@ -1,8 +1,7 @@
-mod feagi_byte_structure_compatible;
 mod feagi_byte_container;
 mod feagi_byte_structure_type;
 mod feagi_serializable;
 
-pub use feagi_byte_container::FeagiByteStructure as FeagiByteStructure;
 pub use feagi_byte_structure_type::FeagiByteStructureType;
-pub use feagi_byte_structure_compatible::FeagiByteStructureCompatible;
+pub use feagi_serializable::FeagiSerializable;
+pub use feagi_byte_container::FeagiByteContainer;
