+++ feagi_data_serialization/src/implementations/mod.rs
@@ -1,2 +1,2 @@
-mod feagi_json;
 mod cortical_mapped_xyzp_neuron_data;
+mod feagi_json;
