+++ feagi_data_serialization/src/lib.rs
@@ -1,4 +1,2 @@
 mod byte_structure;
 mod implementations;
-
-pub use byte_structure::{FeagiByteStructure, FeagiByteStructureCompatible, FeagiByteStructureType};
