+++ feagi_data_structures/src/basic_components/mod.rs
@@ -1,3 +1,3 @@
-//! Macros for defining base data types. Also includes mapping functions to convert between various implementations if needed
+//! Macros for defining base data types. Also includes mapping functions to convert between various implementations-old if needed
 
 mod cartesian;
